+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 import IconButton from '@/components/common/button/IconButton';
-import PostCard from '@/components/common/post-card/PostCard';
-import UserCard from '@/components/common/user-card/UserCard';
+import PostCard from '@/components/common/post-item/post-card/PostCard';
+import UserCard from '@/components/common/post-item/user-card/UserCard';
 import SMore from '@/components/icons/SMore';
 import { postList } from '@/mocks/postList';
 
 export default function Home() {
   return (
     <main className="mt-5 px-4">
-      {postList.map(({ author, ...post }) => (
+      {postList.post.map(({ author, ...post }) => (
         <section key={post.id} className="mb-5">
           <UserCard imageSize="42px" userProfile={author}>
             <IconButton
