+++ src/components/common/post-item/post-card/PostContent.tsx
