+++ src/components/common/post-item/post-card/index.ts
